Adding McDowell paper at 23:00
Pep - this is the McDowell paper, now with all the extra paragraph
spaces and a correction of a mistake I found. This is exactly the same
file as the file of the same name in the "_drafts" folder. It is thus a
paper we could push up to the website.